Refactor: NavBar component - navbar에 있는 로직이 중복이 있어 NavTab으로 로직을 구분하였습니다.
